#741 namespace change by EIP
