+++ src/Nethermind/Nethermind.JsonRpc.Test/ProofCollectorTests.cs
@@ -20,6 +20,7 @@ using Nethermind.Core;
 using Nethermind.Core.Crypto;
 using Nethermind.Core.Test.Builders;
 using Nethermind.Dirichlet.Numerics;
+using Nethermind.JsonRpc.Eip1186;
 using Nethermind.Store;
 using NUnit.Framework;
